pyproject.toml
[project]
name = "wisata-alam"
version = "0.1.0"
description = "Backend for Pariwisata Alam Kaltim"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "bcrypt==4.3.0",
    "python-dotenv>=1.0.1",
    "fastapi[standard]>=0.121.0",
    "python-jose[cryptography]>=3.5.0",
    "ngrok>=1.5.1",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "sqlalchemy>=2.0.44",
]